requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-oauth-toolkit==1.3.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.9
oauthlib==3.1.0
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
psycopg2
gunicorn==20.0.4
